Form Submission - for adding a product

+++ stockMgtService/src/main/java/com/example/stockMgtService/dao/StockRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.stockMgtService.model.Product;
 
-public interface StockRepository extends  JpaRepository<Product, Integer>{
+public interface StockRepository extends  JpaRepository<Product, Long>{
 
 }
